Langserve LCEL/requirements.txt
langchain
python-dotenv
langchain_groq
langchain-groq
langchain_core
langserve
fastapi
uvicorn